+++ .travis.yml
@@ -16,5 +16,3 @@ deploy:
   skip_cleanup: true
   api_key:
     secure: hALpeJGfMD7A+7HfndHiyhPG6ig2CDa1+st/VYgICHhJ++EC8dwn/Cu2ElejmNTZ6sw0gSEpe8zkprCIxHnXvMMuhyYjfr9Ts9NJz/iv8KaVHvDki4JZdx7ceY8LSNFEYEx4Y6Jvo+uzmmMZH5dcYSpd1LLjZRl917AuLUEJU/U=
-after_success:
-  - coveralls
